backend/pyproject.toml
[project]
name = "image-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.11.3",
    "langchain>=0.3.27",
    "langchain-mcp-tools>=0.2.12",
    "langgraph>=0.6.5",
    "websockets>=15.0.1",
    "python-multipart>=0.0.17",
    "pillow>=11.1.0",
    "sse-starlette>=2.2.0",
    "langchain-ollama>=0.3.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.24",
    "alembic>=1.16.5",
    "psycopg2-binary>=2.9.10",
    "psycopg>=3.2.9",
    "aiohttp>=3.12.15",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
]
